requirements.txt
streamlit
langchain
langchain-community
chromadb
sentence-transformers
PyPDF2
python-docx
openpyxl
pandas
numpy
aiohttp
requests
beautifulsoup4
python-dotenv
nemoguardrails
pydantic
tiktoken
psutil